# frontend/src/api/schema/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/department": {
      "post": {
        "tags": [
          "department-controller"
        ],
        "operationId": "upsertDepartment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpsertDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "department-controller"
        ],
        "operationId": "deleteDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/department/query": {
      "get": {
        "tags": [
          "department-controller"
        ],
        "operationId": "queryDepartments",
        "parameters": [
          {
            "name": "departmentQueryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DepartmentQueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Department"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/department/page-query": {
      "get": {
        "tags": [
          "department-controller"
        ],
        "operationId": "pageQueryDepartments",
        "parameters": [
          {
            "name": "pageRequestDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageRequestDto"
            }
          },
          {
            "name": "departmentQueryDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DepartmentQueryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoListDepartment"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DepartmentUpsertDto": {
        "required": [
          "enable",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "enable": {
            "type": "boolean"
          }
        }
      },
...
      "PageRequestDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sortBy": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sortFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortFieldObject"
            }
          }
        }
      },
      "SortFieldObject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "order": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC",
              "DEFAULT"
            ]
          }
        }
      },
...
      "DepartmentQueryDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          },
          "bindState": {
            "type": "string",
            "enum": [
              "BIND",
              "UNBIND",
              "ALL"
            ]
          }
        }
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "enable": {
            "type": "boolean"
          }
        }
      },
      "PageResponseDtoListDepartment": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Department"
            }
          }
        }
      }
    }
  }
}